Delete removed videos
Youtube doesn't immediatly delete a video from the DB, so I have to first check if I can receive some answer and if I can't then the video could be either private or removed. Anyway I set it as removed for tidyness and because some people could keep the video private forever, creating confusion on the local DB.
Anyway this should do the trick.